requirements.txt
httpx[http2]>=0.24.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pyyaml>=6.0
rich>=13.0.0
typer>=0.9.0
jinja2>=3.1.0
aiofiles>=23.0.0
